Rename `MlModel` to `Model`
