+++ lib/src/sql/mod.rs
@@ -38,8 +38,8 @@ pub(crate) mod index;
 pub(crate) mod kind;
 pub(crate) mod language;
 pub(crate) mod limit;
-pub(crate) mod ml_model;
 pub(crate) mod mock;
+pub(crate) mod model;
 pub(crate) mod number;
 pub(crate) mod object;
 pub(crate) mod omit;
