+++ lib/src/sql/statements/define/mod.rs
@@ -4,7 +4,7 @@ mod event;
 mod field;
 mod function;
 mod index;
-mod ml_model;
+mod model;
 mod namespace;
 mod param;
 mod scope;
@@ -18,7 +18,7 @@ pub use event::{event, DefineEventStatement};
 pub use field::{field, DefineFieldStatement};
 pub use function::{function, DefineFunctionStatement};
 pub use index::{index, DefineIndexStatement};
-pub use ml_model::DefineModelStatement;
+pub use model::DefineModelStatement;
 pub use namespace::{namespace, DefineNamespaceStatement};
 use nom::bytes::complete::tag_no_case;
 pub use param::{param, DefineParamStatement};
